First working commit of Arcadia.
The configuration file is coming in the next commit.  With this, the basic
engine implementation should be available.

Arcadia is designed to serve as a base CMS designed for web games.  It
takes a WordPress-like approach to customization, offering actions
for plugin developers to hook on to.  Based on the state of the active
character, we can drive the gaming experience.