First pass of quest completion code.
There are some database changes coming along with this, but of course the
database code isn't live on github yet.  Hmm.

Now we're getting into some real challenges as far as game customization
versus engine flexibility are concerned.  With quests, to make sure we can
be as flexible as possible, we will use eval for state verification.
Without putting in a boatload of custom hacks, like the Twelve Sands quests
had, we're going to rely on the database to be as secure as the code; if
our database server is hosed, we've got bigger problems than allowing eval
to work.