Convert meta_key in db from string to int for performance.
This is mostly a change in the games based on arcadia at this point (i.e.,
you should see Shattered City's changes).  Really though, it means that
we're expecting games to define constants for keys instead of relying on
string values.  We're hitting pages a lot more often, so these database
columns need to be ints.

I think..