Prep for removing quest eval in favour of function/arg calls in meta.
Everything that has a beginning has an end, Neo.

I was always wary of eval, but didn't think there was a better way to
handle things.  With a wider view of what arcadia is supposed to do for
game designers, I think it's important to get rid of arbitrary php code
in the db.  This is evident from the first eval checkin, afaik.

Thank you, camping trip.  Time to think away from a computer.