Allow constant identifiers to be resolved in character_meta.
We don't want magic numbers in the db, for example, and we don't want to
allow arbitrary eval calls.  If we reference a constant in the meta values,
that string won't be present, but it will exist as a defined constant.
Check to see if the constant exists if the actual value doesn't, and if so,
use that one.

It might be possible for an untrusted db to drop some bad values in here,
so maybe we need a set of accepted constants to resolve.