Set global character's meta value on update_character_meta call.
Need to think about how to handle this if the metadata isn't there,
which it isn't guaranteed to be. Probably just set things and have the
post-condition be that regardless of whether or not the meta is there,
we'll set the key because you expect us to.

Also need to handle add_character_meta.