Add character_meta_(int/float) to force type conversion.
I think there's a good argument to be made for putting these in arcadia
instead of just using them in custom game code. A lot of these meta values
are numbers. My code was getting littered with extra parentheses and
intval/floatval calls. In some cases, it was easy to forget the type cast,
and to accidentally try to add or subtract from a string value.