Sample Arcadia project.
This sample shows the basic layout of a project built in Arcadia.
It uses actions, points to a title page, allows character selection,
and other basic things.

I'm going to expand this with calls to game-setting.php and to
start using some of the hooks (that's why tip and credits are there).