Remove default action var and explicit title var. Use microtime for timing.
It's restricting to assume that users will be logged in to set actions.
Title pages should be driven by an action and handled using the engine
instead of dropping out to a title when they're not logged in. At least it
seems like that makes sense.

Also, we don't need to use micro_time for logging. Use the real microtime
with float values.